[REFACTOR] Rename chatrooms to spaces - shard file names
