+++ pandora-server-shard/src/networking/socketio_directory_connector.ts
@@ -3,7 +3,7 @@ const { DIRECTORY_ADDRESS, SERVER_PUBLIC_ADDRESS, SHARD_DEVELOPMENT_MODE, SHARD_
 import { GetLogger, HTTP_HEADER_SHARD_SECRET, HTTP_SOCKET_IO_SHARD_PATH, IShardDirectory, MessageHandler, IDirectoryShard, ConnectionBase, ShardFeature, IDirectoryShardUpdate, RoomId, ShardDirectorySchema, DirectoryShardSchema, IDirectoryShardArgument, IDirectoryShardResult } from 'pandora-common';
 import { connect, Socket } from 'socket.io-client';
 import { CharacterManager } from '../character/characterManager';
-import { RoomManager } from '../room/roomManager';
+import { RoomManager } from '../spaces/spaceManager';
 import { Stop } from '../lifecycle';
 import promClient from 'prom-client';
 import { SocketInterfaceRequest, SocketInterfaceResponse } from 'pandora-common/dist/networking/helpers';
